% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locid.R
\name{check_locid}
\alias{check_locid}
\title{check_locid}
\usage{
check_locid(locid)
}
\arguments{
\item{locid}{Location id}
}
\value{
The message: "\code{locid} is not a valid location id. Please run \code{View(get_locations())} to get a list of plausible location ids. They are listed in the \code{PK_LocID} variable" in the case
where the user has inserted an invalid location id, and in cases where the location id is valid, the following message is returned: "\code{locid} is a valid location id. The location name is \verb{location name}"
}
\description{
Throws an error message if a location id is not part of the location ids in the UNDP database, and a confirmation message showing the location name if the location id is valid.
}
\details{
Run \verb{check_locid(insert locid here)} to check whether a location id is valid (part of the locations in the UNPD website). Running \verb{check_locid(insert locid here)} with a valid id returns a message confirming that the location id is valid and also gives the location name of that particular id. Running the same code with an invalid id returns a message directing the user to run \code{View(get_locations())} in order to get a list of plausible location ids.
You can run \code{getAnywhere(check_locid)} to see how the function is defined.
}
\examples{
\dontrun{
check_locid(12345) ## invalid location id
check_locid(404) ## valid location id
}
}
