% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_age_standard.R
\name{dd_age_standard}
\alias{dd_age_standard}
\title{dd_age_standard}
\usage{
dd_age_standard(data, abridged = TRUE)
}
\arguments{
\item{data}{The data to be harmonized}

\item{abridged}{\code{TRUE/FALSE}}
}
\value{
The DemoData dataset with standard age groups appended to it.
}
\description{
Adds standard age groups (the result of \code{std_age_function()}) in the dataset
}
\details{
Performs a \code{full_join()} operation between the standard age groups dataset (the result of \code{std_age_function()}) and the data to be harmonized,
using \code{AgeStart},\code{AgeEnd},\code{AgeLabel},\code{AgeSpan} as keys. You can run \code{getAnywhere(dd_age_standard)} to view the definition of this function.
}
\examples{
\dontrun{
df <- firstages_compute_births_df
df <- dd_age_standard(df)
}
}
