% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_append_tcs_cas.R
\name{dd_append_tcs_cas}
\alias{dd_append_tcs_cas}
\title{dd_append_tcs_cas}
\usage{
dd_append_tcs_cas(indata, type, tcs_data, ind)
}
\arguments{
\item{indata}{Births / Deaths data that is already harmonized}

\item{type}{births or deaths}

\item{tcs_data}{Total counts by sex data}

\item{ind}{159 (where type == "births) or 188 (where type == "deaths)}
}
\value{
A dataset that has both Total counts by sex data and Counts by age and sex
}
\description{
dd_append_tcs_cas
}
