% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_births_abridged2single_1014.R
\name{dd_births_abridged2single_1014}
\alias{dd_births_abridged2single_1014}
\title{dd_births_abridged2single_1014}
\usage{
dd_births_abridged2single_1014(indata, births10_14)
}
\arguments{
\item{indata}{The data to be harmonized}

\item{births10_14}{The data value of age label "10-14"}
}
\value{
A dataset with ages 10-14 distributed across single ages.
}
\description{
Distribute births to ages 10-14 across single ages. Arbitrary distribution of births across ages 10-14 for now. Need to update with Kirill's regression
}
\examples{
\dontrun{
df <- abridged2single_1014_df
df <- dd_births_abridged2single_1014(df, 4)
}
}
