% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_distribute_unknowns.R
\name{dd_distribute_unknowns}
\alias{dd_distribute_unknowns}
\title{dd_distribute_unknowns}
\usage{
dd_distribute_unknowns(data)
}
\arguments{
\item{data}{The data to be harmonised}
}
\value{
A dataset without unknowns
}
\description{
Distribute Unknowns over age
}
\details{
You can run \code{getAnywhere(dd_distribute_unknowns)} to see how the function is defined.
}
\examples{
\dontrun{
df <- vitals_std_valid_sample
df <- dd_distribute_unknowns(df)
}
}
