% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_extract_single.R
\name{dd_extract_single}
\alias{dd_extract_single}
\title{dd_extract_single}
\usage{
dd_extract_single(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with single-year of age records only
}
\description{
Extract records that refer to single-year of age. Sometimes a series identified as abridged in DemoData includes records by single year of age We extract these for use on a single year series
}
\examples{
\dontrun{
df <- data.frame()
}
}
