% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_firstages_compute.R
\name{dd_firstages_compute}
\alias{dd_firstages_compute}
\title{dd_firstages_compute}
\usage{
dd_firstages_compute(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with ages \code{0-1}, \code{1-4} and \code{0-4} reconciled.
}
\description{
Reconcile ages 0-1, 1-4 and 0-4
}
\details{
If \code{0-4} is missing and \code{0-1} and \code{1-4} are present, then sum both to get \code{0-4}.
You can run \code{getAnywhere(dd_firstages_compute)} to view the definition of this function.
}
\examples{
\dontrun{
df <- firstages_compute_df
df <- dd_firstages_compute(df)
}
}
