% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_firstages_compute_births.R
\name{dd_firstages_compute_births}
\alias{dd_firstages_compute_births}
\title{dd_firstages_compute_births}
\usage{
dd_firstages_compute_births(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with early age groups reconciled e.g 0-14 becomes 10-14
}
\description{
Reconcile early age groups for births.
}
\details{
In cases where for example 10_14 is missing and 0_14 is present, convert the latter to the former.
Another case is if 0-4 is missing and 0-1 and 1-4 are present, then sum both to get 0-4.
You can run \code{getAnywhere(dd_firstages_compute_births)} to view the definition of this function.
}
\examples{
\dontrun{
df <- firstages_compute_births_df
df <- dd_firstages_compute_births(df)
}
}
