% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_multiple_totals.R
\name{dd_multiple_totals}
\alias{dd_multiple_totals}
\title{dd_multiple_totals}
\usage{
dd_multiple_totals(data)
}
\arguments{
\item{data}{The data to be harmonised}
}
\value{
A data frame with only one \code{Total} age label per id, SexId and indicator id.
}
\description{
If there exists more than one \code{Total} age label and there is one that is equal to the computed total, drop the others and keep this one.
}
