% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_oag2closed.R
\name{dd_oag2closed}
\alias{dd_oag2closed}
\title{dd_oag2closed}
\usage{
dd_oag2closed(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with closed age groups derived from multiple open age groups.
}
\description{
Compute associated closed age group values from multiple open-age groups
}
\details{
You can run \code{getAnywhere(dd_oag2closed)} to view the definition of this function.
}
