% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_oag_multiple.R
\name{dd_oag_multiple}
\alias{dd_oag_multiple}
\title{dd_oag_multiple}
\usage{
dd_oag_multiple(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
\code{TRUE/FALSE}
}
\description{
Checks whether the series has more than one open age group.
}
\details{
An open age group is identified as one where the \code{AgeSpan} == -1 and \code{AgeStart} != 0.
You can run \code{getAnywhere(dd_oag_multiple)} to view the definition of this function.
}
\examples{
\dontrun{
dd_oag_multiple(oag_multiple_df)
}
}
