% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_rank_id.R
\name{dd_rank_id}
\alias{dd_rank_id}
\title{dd_rank_id}
\usage{
dd_rank_id(indata)
}
\arguments{
\item{indata}{The data to be harmonized}
}
\value{
A dataset with only one id per country and year
}
\description{
Ranking id; For each country- year, we rank the available ids for population counts.
}
\details{
See Part 3 of the \href{https://shelmith-kariuki.github.io/rddharmony/articles/Harmonization_Workflow.html}{Harmonization Workflow article} for details about this function.
}
