% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_rank_id_vitals.R
\name{dd_rank_id_vitals}
\alias{dd_rank_id_vitals}
\title{dd_rank_id_vitals}
\usage{
dd_rank_id_vitals(indata)
}
\arguments{
\item{indata}{The data to be harmonised}
}
\value{
A dataset with only one id per country and year.
}
\description{
Ranking id; For each country- year, we rank the available ids for counts of vital events.
}
\details{
See Part 3 of the \href{https://shelmith-kariuki.github.io/rddharmony/articles/Harmonization_Workflow.html}{Harmonization Workflow article} for details about this function.
}
\examples{
\dontrun{
df <- vitals_std_valid_sample2
df <- dd_rank_id_vitals(df)
}
}
