% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_validate_totals_over_age.R
\name{dd_validate_totals_over_age}
\alias{dd_validate_totals_over_age}
\title{dd_validate_totals_over_age}
\usage{
dd_validate_totals_over_age(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with reconciled \code{Totals} i.e. reported totals == computed totals
}
\description{
Reconcile \code{Totals}, in cases where reported values are different from calculated values, after having standardized age series.
If the computed total is greater than reported total, then replace reported with computed. If computed is less than reported, then add difference to "Unknown" age
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/dd_validate_totals_over_age.html}{"Validating totals over age" vignette} for more details about this function.
}
\examples{
\dontrun{
df <- vitals_std_full_sample
df <- dd_validate_totals_over_age(df)
}
}
