% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_validate_totals_over_sex_new.R
\name{dd_validate_totals_over_sex_new}
\alias{dd_validate_totals_over_sex_new}
\title{dd_validate_totals_over_sex_new}
\usage{
dd_validate_totals_over_sex_new(data)
}
\arguments{
\item{data}{Data to be harmonized}
}
\value{
A dataset where Both sexes == Female + Male for each id and SexId.
}
\description{
Compute total over sex as max of reported total or sum over sex, compute any NA values based on differences,
redistribute SexID = 0 across males and females then validate totals over sex.
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/dd_validate_totals_over_sex.html}{"Validating totals over sex" vignette} for more details about this function.
}
\examples{
\dontrun{
df <- validate_totals_over_sex_pop
df <- dd_validate_totals_over_sex_new(df)
}
}
