% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDextract_PopCounts.R
\name{DDextract_PopCounts}
\alias{DDextract_PopCounts}
\title{DDextract_PopCounts}
\usage{
DDextract_PopCounts(
  locid,
  process = c("census", "estimate", "register"),
  start_year,
  end_year,
  locType = "whole",
  DataSourceShortName = NULL,
  DataSourceYear = NULL
)
}
\arguments{
\item{locid}{The Location Id for each country. Run \code{DemoTools::get_locations()} to get information about available locations. The ids are indicated in the \code{PK_LocID} variable.}

\item{process}{The data collection process i.e. Census (census), Estimate (estimate) or Register (register).}

\item{start_year}{The minimum year for the data available for each country}

\item{end_year}{The maximum year for the data available for each country}

\item{locType}{The location type for which data are extracted i.e. Whole, Urban, Rural}

\item{DataSourceShortName}{NULL.}

\item{DataSourceYear}{NULL.}
}
\value{
A dataset showing population counts for each location, process, year, sex and age label.
}
\description{
Extract population counts by age and sex for a given country and time period can choose process = "census", "estimate", or "register"
}
\examples{
\dontrun{
dd_extract <- DDextract_PopCounts(404, #Kenya
                                   process = c("census","estimate","register"),
                                   1950,
                                   2050,
                                   DataSourceShortName = NULL,
                                   DataSourceYear = NULL)
}
}
