% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDextract_PopTotals.R
\name{DDextract_PopTotals}
\alias{DDextract_PopTotals}
\title{DDextract_PopTotals}
\usage{
DDextract_PopTotals(
  locid,
  process = c("census", "estimate", "register"),
  start_year,
  end_year,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)
}
\arguments{
\item{locid}{The Location Id for each country. Run \code{DemoTools::get_locations()} to get information about available locations. The ids are indicated in the \code{PK_LocID} variable.}

\item{process}{The data collection process i.e. Census (census), Estimate (estimate) or Register (register).}

\item{start_year}{The minimum year for the data available for each country}

\item{end_year}{The maximum year for the data available for each country}

\item{DataSourceShortName}{NULL.}

\item{DataSourceYear}{NULL.}
}
\value{
A data.frame showing population counts for each location, process, year, location type and sex.
}
\description{
Extract population totals, by sex and location type, for a given country and time period can choose process = "census", "estimate", or "register"
}
\examples{
\dontrun{
dd_extract <- DDextract_PopTotals(404, #Kenya
                                   process = c("census","estimate","register"),
                                   1950,
                                   2050,
                                   DataSourceShortName = NULL,
                                   DataSourceYear = NULL)
}
}
