% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDextract_VitalCounts.R
\name{DDextract_VitalCounts}
\alias{DDextract_VitalCounts}
\title{DDextract_VitalCounts}
\usage{
DDextract_VitalCounts(
  locid,
  type = c("births", "deaths"),
  process = c("census", "vr"),
  start_year,
  end_year,
  DataSourceShortName = NULL,
  DataSourceYear = NULL
)
}
\arguments{
\item{locid}{The Location Id for each country. Run \code{DemoTools::get_locations()} to get information about available locations. The ids are indicated in the \code{PK_LocID} variable.}

\item{type}{The type of data to be pulled i.e. births/deaths.}

\item{process}{The data collection process i.e. Census (census) or Vital Registrations (vr).}

\item{start_year}{The minimum year for the data to be extracted.}

\item{end_year}{The maximum year for the data to be extracted.}

\item{DataSourceShortName}{NULL.}

\item{DataSourceYear}{NULL.}
}
\value{
A dataset showing counts for each location, type of data (births or deaths), process, year, sex and age label.
}
\description{
Extracts births and/or deaths data for a given country and time period from the UNDP portal.
}
\details{
This function extracts data using \href{https://timriffe.github.io/DDSQLtools/reference/get_recorddata.html}{get_recorddata()}.
The function returns \code{NULL} in cases where the data is not available and a dataframe of 93 variables where it exists.
In cases where the user inserts an invalid location id, a message is returned prompting them to run \code{View(get_locations())} to get a list of plausible location ids. See
\link{check_locid} for more details on this. You can run \code{getAnywhere(DDextract_VitalCounts)} to see how the function is defined.
}
\examples{
\dontrun{
# Extracting births data
births_df <- DDextract_VitalCounts(404, #Kenya
                                   type = c("births"),
                                   process = c("census","vr"),
                                   1950,
                                   2050,
                                   DataSourceShortName = NULL,
                                   DataSourceYear = NULL)

# Extracting deaths data
deaths_df <- DDextract_VitalCounts(364, #Iran
                                   type = c("deaths"),
                                   process = c("census","vr"),
                                   1950,
                                   2050,
                                   DataSourceShortName = NULL,
                                   DataSourceYear = NULL)
}
}
