% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_Pop1.R
\name{DDharmonize_Pop1}
\alias{DDharmonize_Pop1}
\title{DDharmonize_Pop1}
\usage{
DDharmonize_Pop1(indata)
}
\arguments{
\item{indata}{The data to be harmonised.}
}
\value{
A data frame that contains harmonized single year of age counts and closed by an open age group.
}
\description{
This wrapper takes population1 indicators from DemoData and Standardizes/harmonizes them.
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/DDharmonize_1YearCounts.html}{"Harmonizing one-year counts" vignette} for more details about this function.
}
\examples{
\dontrun{
df <- pop1_df
df <- DDharmonize_Pop1(df)
}
}
