% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_Pop5.R
\name{DDharmonize_Pop5}
\alias{DDharmonize_Pop5}
\title{DDharmonize_Pop5}
\usage{
DDharmonize_Pop5(indata)
}
\arguments{
\item{indata}{The data to be harmonised}
}
\value{
A data frame that contains standard abridged age groups 0, 1-4, 0-4, 5-9, 10-14 ..... up to the open age group
}
\description{
This wrapper takes Population5 indicators from DemoData and Standardizes/harmonizes them.
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/DDharmonize_5YearCounts.html}{"Harmonizing five-year counts" vignette} for more details about this function.
}
