% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_validate_PopCounts.R
\name{DDharmonize_validate_PopCounts}
\alias{DDharmonize_validate_PopCounts}
\title{DDharmonize_validate_PopCounts}
\usage{
DDharmonize_validate_PopCounts(
  locid,
  dd_extract = NULL,
  times,
  process = c("census", "estimate", "register"),
  locType = "whole",
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)
}
\arguments{
\item{locid}{Location id}

\item{dd_extract}{A data frame extracted from DemoData using the DDextract_PopCounts() function. If NULL then the function will query DemoData for a new extract.}

\item{times}{The period of the data to be extracted. You can extract one year data e.g times = 2020 or a longer period of time e.g times = c(1950, 2020).}

\item{process}{The process through which the data was obtained from various sources i.e either via census , register or estimates. By default, the function pulls data obtained through both of these processes.}

\item{locType}{The location type for which data are extracted i.e. Whole, Urban, Rural}

\item{return_unique_ref_period}{TRUE / FALSE. Specifies whether the data to be returned should contain one unique id (return_unique_ref_period == TRUE) or several ids (return_unique_ref_period == FALSE) per time label.
ids are a unique identifier for each unique set of records based on \code{LocID}, \code{LocName}, \code{DataProcess}, \code{ReferencePeriod}, \code{DataSourceName}, \code{StatisticalConceptName}, \code{DataTypeName} and \code{DataReliabilityName}.
The definitions of these variables are provided later in this article.}

\item{DataSourceShortName}{NULL}

\item{DataSourceYear}{NULL}

\item{retainKeys}{TRUE / FALSE. Specifies whether only a few (retainKeys == FALSE) or all (retainKeys == TRUE) variables should be retained in the output.}

\item{server}{"https://popdiv.dfs.un.org/DemoData/api/"}
}
\value{
A harmonized dataset containing population counts
}
\description{
This script implements a work flow for census population counts extracting from DemoData, harmonizing age groups,
identifying full series, selecting preferred series, validating totals and by sex and eventually
including key fields in the function output.
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/DDharmonize_validate_PopCounts.html}{DDharmonize_validate_PopCounts} vignette for more details about this function.
}
\examples{
\dontrun{
sweden_df <- DDharmonize_validate_PopCounts(locid = 752,
                                              times = c(1950,2020),
                                              process = c("census", "estimate", "register"),
                                              return_unique_ref_period = TRUE,
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")
}
}
