% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_Vitals1.R
\name{DDharmonize_Vitals1}
\alias{DDharmonize_Vitals1}
\title{DDharmonize_Vitals1}
\usage{
DDharmonize_Vitals1(indata)
}
\arguments{
\item{indata}{The data to be harmonised}
}
\value{
A data frame that contains harmonized single year of age counts and closed by an open age group.
}
\description{
This wrapper takes vital counts from census and registers from DemoData and Standardizes/harmonizes them.
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/DDharmonize_1YearCounts.html}{"Harmonizing one-year counts" vignette} for more details about this function.
}
