% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_Vitals5.R
\name{DDharmonize_Vitals5}
\alias{DDharmonize_Vitals5}
\title{DDharmonize_Vitals5}
\usage{
DDharmonize_Vitals5(indata, type = c("births", "deaths"))
}
\arguments{
\item{indata}{The data to be harmonised}

\item{type}{births / deaths}
}
\value{
A data frame that contains harmonized counts for five-year age groups and closed by an open age group.
}
\description{
This wrapper takes five-year vital counts from census and registers from the UNDP database and standardizes/harmonizes them.
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/DDharmonize_5YearCounts.html}{"Harmonizing five-year counts" vignette} for more details about this function.
}
\examples{
\dontrun{
df <- vitals5_df
df <- DDharmonize_Vitals5(df, type = "births")
}
}
