% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dropped_ids.R
\name{print_dropped_ids}
\alias{print_dropped_ids}
\title{print_dropped_ids}
\usage{
print_dropped_ids(indata_before, indata_after)
}
\arguments{
\item{indata_before}{vitals_std_all}

\item{indata_after}{vitals_std_full}
}
\value{
ids that have been dropped because they result in data that is not full
}
\description{
Print ids that have been dropped, i.e those that are not full series
}
\keyword{internal}
