% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_age_function.R
\name{std_age_function}
\alias{std_age_function}
\title{std_age_function}
\usage{
std_age_function()
}
\value{
A dataset with standard age groups
}
\description{
Creates a data frame with all of the standard age groups we would expect for abridged and complete series (up to 130+ open age group)
}
\details{
You can run \code{getAnywhere(std_age_function)} to view the definition of this function.
}
\examples{
\dontrun{
std_ages <- std_age_function()
}
}
